add authentication

+++ notes-backend/src/note/note.service.ts
@@ -1,10 +1,9 @@
 import { Injectable } from '@nestjs/common';
-// import { CreateNoteDto } from './dto/create-note.dto';
-// import { UpdateNoteDto } from './dto/update-note.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Note } from './entities/note.entity';
 import { DeleteResult, Repository } from 'typeorm';
 import { UUID } from 'crypto';
+import { CreateNoteDto } from './dto/create-note.dto';
 
 @Injectable()
 export class NoteService {
@@ -12,9 +11,13 @@ export class NoteService {
     @InjectRepository(Note)
     private repository: Repository<Note>,
   ) {}
-  // async create(createNoteDto: CreateNoteDto) {
-  //   return 'This action adds a new note';
-  // }
+  create(createNoteDto: CreateNoteDto, userId: UUID) {
+    return this.repository.create({
+      title: createNoteDto.title,
+      content: createNoteDto.content,
+      userId: userId,
+    });
+  }
 
   async findAll(): Promise<Note[]> {
     return await this.repository.find();
@@ -24,11 +27,7 @@ export class NoteService {
     return await this.repository.findOneBy({ id });
   }
 
-  // async update(id: number, updateNoteDto: UpdateNoteDto) {
-  //   return `This action updates a #${id} note`;
-  // }
-
-  async delete(id: UUID): Promise<DeleteResult> {
-    return await this.repository.delete(id);
+  async delete(id: UUID, userId: UUID): Promise<DeleteResult> {
+    return await this.repository.delete({ id, userId });
   }
 }
